+++ R/bbwp_farm_score.R
@@ -8,17 +8,17 @@
 #' @param S_BBWP_PSW (numeric) the scoring index for lowering P emission to surface water (risk times impact) for each field
 #' @param S_BBWP_NUE (numeric) the scoring index to use N and P inputs efficiently for each field
 #' @param S_BBWP_WB (numeric) the scoring index to buffer and store water and efficiently use water for plant growth for each field
-#' @param D_AREA (numeric) the area of the field (\ m2 or \ ha) 
+#' @param B_AREA (numeric) the area of the field (m2) 
 #'   
 #' @import data.table
 #'
 #' @export
 # calculate the opportunities for a set of fields
-bbwp_farm_score <- function(S_BBWP_TOT,S_BBWP_NGW,S_BBWP_NSW,S_BBWP_PSW,S_BBWP_NUE,S_BBWP_WB, D_AREA){
+bbwp_farm_score <- function(S_BBWP_TOT,S_BBWP_NGW,S_BBWP_NSW,S_BBWP_PSW,S_BBWP_NUE,S_BBWP_WB, B_AREA){
 
   # check length of the inputs
   arg.length <- max(length(S_BBWP_TOT),length(S_BBWP_NGW),length(S_BBWP_NSW),length(S_BBWP_PSW),
-                    length(S_BBWP_NUE),length(S_BBWP_WB),length(D_AREA))
+                    length(S_BBWP_NUE),length(S_BBWP_WB),length(B_AREA))
   
   # check inputs
   checkmate::assert_numeric(S_BBWP_TOT, lower = 0, upper = 100, len = arg.length)
@@ -27,7 +27,7 @@ bbwp_farm_score <- function(S_BBWP_TOT,S_BBWP_NGW,S_BBWP_NSW,S_BBWP_PSW,S_BBWP_N
   checkmate::assert_numeric(S_BBWP_PSW, lower = 0, upper = 100, len = arg.length)
   checkmate::assert_numeric(S_BBWP_NUE, lower = 0, upper = 100, len = arg.length)
   checkmate::assert_numeric(S_BBWP_WB, lower = 0, upper = 100, len = arg.length)
-  checkmate::assert_numeric(D_AREA, lower = 0, upper = 50000, len = arg.length)
+  checkmate::assert_numeric(B_AREA, lower = 0, upper = 50000, len = arg.length)
   
   # collect data in one data.table
   dt <- data.table(id = 1:arg.length,
@@ -37,14 +37,14 @@ bbwp_farm_score <- function(S_BBWP_TOT,S_BBWP_NGW,S_BBWP_NSW,S_BBWP_PSW,S_BBWP_N
                    S_BBWP_NUE = S_BBWP_NUE,
                    S_BBWP_WB = S_BBWP_WB,
                    S_BBWP_TOT = S_BBWP_TOT,
-                   D_AREA = D_AREA
+                   B_AREA = B_AREA
                   )
   
   # columns with the score of the opportunity indexes
   cols <- c('S_BBWP_TOT','S_BBWP_NGW','S_BBWP_NSW','S_BBWP_PSW','S_BBWP_NUE','S_BBWP_WB')
   
   # calculate area weigthed sum of the field indices
-  dt <- dt[,lapply(.SD, stats::weighted.mean, w = D_AREA), .SDcols = cols]
+  dt <- dt[,lapply(.SD, stats::weighted.mean, w = B_AREA), .SDcols = cols]
   
   # Round the values
   dt<- dt[, lapply(.SD, round, digits = 0)]