+++ R/bbwp_main.R
@@ -31,11 +31,11 @@
 #' @param A_P_SG (numeric) The P-saturation index (\%)
 #' @param D_SA_W (numeric) The wet perimeter index of the field, fraction that field is surrounded by water
 #' @param D_RO_R (numeric) The risk that surface water runs off the parcel
-#' @param D_AREA (numeric) the area of the field (\ m2 or \ ha) 
+#' @param B_AREA (numeric) the area of the field (m2) 
 #' @param M_DRAIN (boolean) is there tube drainage present in the field
 #' @param LSW (data.table) The surface water polygon for catchment or polder (NULL if not available, lat/lon should be provided)
-#' @param lat (numeric) Latitude of the field (required if no LSW is submitted)
-#' @param lon (numeric) Longitude of the field (required if no LSW is submitted)
+#' @param a_lat (numeric) Latitude of the field (required if no LSW is submitted)
+#' @param a_lon (numeric) Longitude of the field (required if no LSW is submitted)
 #' @param measures (data.table) the measures planned / done per fields
 #' @param sector (string) a vector with the farm type given the agricultural sector (options: options: 'diary', 'arable', 'tree_nursery', 'bulbs')
 #' @param output (string) a vector specifying the output type of the function. Options: scores, measures 
@@ -51,8 +51,8 @@ bbwp <- function(B_SOILTYPE_AGR, B_LU_BRP, B_LU_BBWP,B_GWL_CLASS, B_SC_WENR, B_H
                  B_AER_CBS,
                  A_CLAY_MI, A_SAND_MI, A_SILT_MI, A_SOM_LOI, A_N_RT,A_FE_OX, A_AL_OX, A_P_CC, A_P_AL, A_P_WA, A_P_SG,
                  B_GWP, B_AREA_DROUGHT, B_CT_PSW, B_CT_NSW,B_CT_PSW_MAX = 0.5, B_CT_NSW_MAX = 5.0, 
-                 D_SA_W, D_RO_R, D_AREA, 
-                 M_DRAIN, LSW, lat = NULL, lon = NULL,
+                 D_SA_W, D_RO_R, B_AREA, 
+                 M_DRAIN, LSW, a_lat = NULL, a_lon = NULL,
                  measures, sector,output = 'scores'){
   
   # add visual binding
@@ -97,8 +97,8 @@ bbwp <- function(B_SOILTYPE_AGR, B_LU_BRP, B_LU_BBWP,B_GWL_CLASS, B_SC_WENR, B_H
                               D_SA_W = D_SA_W, 
                               D_RO_R =  D_RO_R, 
                               LSW = LSW, 
-                              lat = lat, 
-                              lon = lon)
+                              a_lat = a_lat, 
+                              a_lon = a_lon)
   
   # Aggregate BBWP risk indicators into five indicators
   dt.ind <- bbwp_field_indicators(D_NGW_SCR = dt$ngw_scr,
@@ -164,7 +164,7 @@ bbwp <- function(B_SOILTYPE_AGR, B_LU_BRP, B_LU_BBWP,B_GWL_CLASS, B_SC_WENR, B_H
                              S_BBWP_PSW = dt.fields$S_BBWP_PSW,
                              S_BBWP_NUE = dt.fields$S_BBWP_NUE,
                              S_BBWP_WB = dt.fields$S_BBWP_WB,
-                             D_AREA = D_AREA)
+                             B_AREA = B_AREA)
                               
   # return output when preferred measures are requested
   if(output == 'measures'){