+++ R/er_meas_scores.R
@@ -16,11 +16,11 @@
 er_meas_score <- function(B_SOILTYPE_AGR, B_LU_BRP,B_LU_BBWP,B_AER_CBS, measures, sector){
   
   # add visual bindings
-  eco_id = type = fr_area = D_AREA = id = er_urgency = NULL
+  eco_id = type = fr_area = id = er_urgency = NULL
   fsector = fdairy = dairy = farable = arable = ftree_nursery = tree_nursery = fbulbs = bulbs = NULL
   crop_cat1 = crop_cat2 = crop_cat3 = crop_cat4 = crop_cat5 = crop_cat6 = crop_cat7 = crop_cat8 = crop_cat9 = NULL
   soiltype = peat = clay = sand = silt = loess = NULL
-  patterns = indicator = erscore = urgency = reward = value = NULL
+  patterns = indicator = erscore = urgency = S_ER_REWARD = value = NULL
   total = biodiversity = climate = landscape = soil = water = oid = NULL
   
   # reformat B_AER_CBS
@@ -52,8 +52,7 @@ er_meas_score <- function(B_SOILTYPE_AGR, B_LU_BRP,B_LU_BBWP,B_AER_CBS, measures
     B_SOILTYPE_AGR = B_SOILTYPE_AGR,
     B_LU_BRP = B_LU_BRP,
     B_LU_BBWP = B_LU_BBWP,
-    B_AER_CBS = B_AER_CBS,
-    D_AREA = D_AREA
+    B_AER_CBS = B_AER_CBS
   )
   
   # merge all measures to the given fields
@@ -135,7 +134,7 @@ er_meas_score <- function(B_SOILTYPE_AGR, B_LU_BRP,B_LU_BBWP,B_AER_CBS, measures
   dt.field <- dt2[,lapply(.SD,sum), .SDcols = cols, by = 'id']
     
   # calculate total reward per field (euro / ha)
-  dt.reward <- dt[indicator == 'profit',list(reward = sum(value)),by = 'id']
+  dt.reward <- dt[indicator == 'profit',list(S_ER_REWARD = sum(value)),by = 'id']
   
   # add reward to the field
   dt.field <- merge(dt.field,dt.reward,by='id')