+++ tests/testthat/test-er_farm_score.R
@@ -6,8 +6,8 @@
   # S_ER_CLIMATE = seq(0, 10, 2.5)
   # S_ER_BIO = seq(0, 10, 2.5)
   # S_ER_LANDSCAPE = seq(0, 10, 2.5)
-  # reward = c(250,4650,213,15,680)
-  # D_AREA = 1
+  # S_ER_REWARD = c(250,4650,213,15,680)
+  # B_AREA = 1
 
 
 test_that("check er_farm_score", {
@@ -19,8 +19,8 @@ test_that("check er_farm_score", {
       S_ER_CLIMATE = seq(0, 10, 2.5),
       S_ER_BIODIVERSITY = seq(0, 10, 2.5),
       S_ER_LANDSCAPE = seq(0, 10, 2.5),
-      reward = c(250,4650,213,15,680),
-      D_AREA = rep(1,5)
+      S_ER_REWARD = c(250,4650,213,15,680),
+      B_AREA = rep(1,5)
     ),
     expected = data.table(
       S_ER_TOT = 5,
@@ -29,7 +29,7 @@ test_that("check er_farm_score", {
       S_ER_CLIMATE = 5,
       S_ER_BIODIVERSITY = 5,
       S_ER_LANDSCAPE = 5,
-      reward = 1162
+      S_ER_REWARD = 1162
     ),
     tolerance = 0.01)
 })
@@ -45,8 +45,8 @@ test_that("check er_farm_score", {
       S_ER_CLIMATE = seq(0, 10, 2.5),
       S_ER_BIODIVERSITY = seq(0, 10, 2.5),
       S_ER_LANDSCAPE = seq(0, 10, 2.5),
-      reward = c(250,4650,213,15,680),
-      D_AREA = c(15,0.8,45,95,12)
+      S_ER_REWARD = c(250,4650,213,15,680),
+      B_AREA = c(15,0.8,45,95,12)
     ),
     expected = data.table(
       S_ER_TOT = 63,
@@ -55,7 +55,8 @@ test_that("check er_farm_score", {
       S_ER_CLIMATE = 6,
       S_ER_BIODIVERSITY = 6,
       S_ER_LANDSCAPE = 6,
-      reward = 159
+      S_ER_REWARD = 159
     ),
     tolerance = 0.01)
 })
+